requirements.txt
streamlit
opencv-python
easyocr
imutils
pillow
